Remove unnecessary repo name
Removing the repo name allows to run the pipeline in a repo with another name than the original. This prevents to change or remove 'antora/antora-ui-default' as it is now name agnostic, e.g., when someone forks the repo and runs the pipeline. Without this change those jobs with 'antora/antora-ui-default' would not be included/excluded correctly.